backend/requirements-windows.txt
# Windows-compatible requirements
# Use this if you're on Windows and having build issues

# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database - Windows-friendly versions
sqlalchemy==2.0.27
alembic==1.13.1
# asyncpg==0.29.0  # Comment out - install separately if needed
# psycopg2-binary==2.9.9  # Comment out - install separately

# Validation
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
supabase==2.3.4

# Caching & Background Tasks
redis==5.0.1
celery==5.3.6
flower==2.0.1

# HTTP Client
httpx==0.25.2  # Compatible with supabase
aiohttp==3.9.3

# Utilities
python-dotenv==1.0.1
pendulum==3.0.0
loguru==0.7.2

# Monitoring
sentry-sdk[fastapi]==1.40.5

# AI & NLP
openai==1.12.0
# tiktoken==0.6.0  # Comment out - install separately

# Cloud Storage
boto3==1.34.44

# Email
resend==0.7.2
